feat: adiciona GitHubService e interfaces para manipulação de PRs
