+++ .github/workflows/ci.yml
@@ -22,6 +22,7 @@ jobs:
           OPENAI_API_KEY: ${{ secrets.OPENAI_API_KEY }}
           OPENAI_ASSISTANT_ID: ${{ secrets.OPENAI_ASSISTANT_ID }}
           exclude: '**/*.json, **/*.md, **/*.map, **/*.lock.json, dist/**/*.js'
+          MYSQL_URL: ${{ secrets.MYSQL_URL }}
 
       # Use the output from the `hello` step
       - name: Get the output
