+++ src/lib/github/index.ts
@@ -1,0 +1,8 @@
+import { GitHubService } from './github.service';
+
+export * from './github.interface';
+export type { GitHubService };
+
+export function createGithubService() {
+  return new GitHubService();
+}
